flow works fully but need to fix error in subgraph deleting revoked domains to create properly, subgraph is not working ATM so blocked on that
